+++ src/collective/collectionfilter/baseviews.py
@@ -2,25 +2,28 @@
 import json
 
 from plone import api
+from plone.memoize import instance
 
 from collective.collectionfilter import PLONE_VERSION
 from collective.collectionfilter.filteritems import get_filter_items
+from collective.collectionfilter.interfaces import IGroupByCriteria
 from collective.collectionfilter.query import make_query
 from collective.collectionfilter.filteritems import get_section_filter_items
 from collective.collectionfilter.utils import base_query
+from collective.collectionfilter.utils import safe_iterable
 from collective.collectionfilter.utils import safe_decode
 from collective.collectionfilter.utils import safe_encode
 from collective.collectionfilter.vocabularies import TEXT_IDX
 from plone.api.portal import get_registry_record as getrec
 from plone.app.contenttypes.behaviors.collection import ICollection
 from Acquisition import aq_inner
-# from collective.collectionfilter import PLONE_VERSION
 from plone.app.uuid.utils import uuidToCatalogBrain
 from plone.app.uuid.utils import uuidToObject
 from plone.i18n.normalizer.interfaces import IIDNormalizer
 from Products.CMFPlone.utils import safe_unicode
 from six.moves.urllib.parse import urlencode
 from zope.component import queryUtility
+from zope.component import getUtility
 from zope.i18n import translate
 from Products.CMFPlone.utils import get_top_request
 
@@ -119,6 +122,9 @@ class BaseFilterView(BaseView):
         else:
             return 'checkbox'
 
+    # results is called twice inside the template in view/available and view/results.  But its expensive so we cache it
+    # but just the the lifetime of the view
+    @instance.memoize
     def results(self):
         results = get_filter_items(
             target_collection=self.settings.target_collection,
@@ -167,6 +173,22 @@ class BaseSectionView(BaseView):
         )
         return results
 
+    @property
+    def is_available(self):
+        if not self.settings.hide_if_empty:
+            return True
+
+        if self.settings.narrow_down:
+            groupby_criteria = getUtility(IGroupByCriteria).groupby
+            idx = groupby_criteria[self.settings.group_by]['index']
+            request_params = safe_decode(self.top_request.form)
+            current_idx_value = safe_iterable(request_params.get(idx))
+            if current_idx_value:
+                return True
+
+        results = self.results()
+        return not (results is None or len(results) <= 2)  # 2 becayse we include "All"
+
 
 class BaseSearchView(BaseView):
 
@@ -208,6 +230,10 @@ class BaseSearchView(BaseView):
             '?' + query_param if query_param else None
         ] if it])
         return ajax_url
+
+    @property
+    def is_available(self):
+        return True
 
 
 if HAS_GEOLOCATION: