+++ src/collective/collectionfilter/portlets/collectionfilter.py
@@ -28,6 +28,7 @@ class Assignment(base.Assignment):
     narrow_down = False
     view_name = None
     content_selector = '#content-core'
+    hide_if_empty = False
     # list_scaling = None
 
     def __init__(
@@ -42,6 +43,7 @@ class Assignment(base.Assignment):
         narrow_down=False,
         view_name=None,
         content_selector='#content-core',
+        hide_if_empty=False,
         # list_scaling=None
     ):
         self.header = header
@@ -54,6 +56,7 @@ class Assignment(base.Assignment):
         self.narrow_down = narrow_down
         self.view_name = view_name
         self.content_selector = content_selector
+        self.hide_if_empty = hide_if_empty
         # self.list_scaling = list_scaling
 
     @property