+++ src/collective/collectionfilter/resources/collectionfilter.js
@@ -2,7 +2,7 @@ define([
     'jquery',
     'pat-base',
     'mockup-patterns-contentloader'
-], function($, Base, contentloader) {
+], function ($, Base, contentloader) {
 
     var CollectionFilter = Base.extend({
         name: 'collectionfilter',
@@ -19,7 +19,7 @@ define([
         _initmap_cycles: 2,
         _zoomed: false,
 
-        init: function() {
+        init: function () {
             this.$el.unbind('collectionfilter:reload');
             this.$el.on('collectionfilter:reload', function (e, data) {
                 if ((data.noReloadSearch && this.$el.hasClass('collectionSearch')) ||
@@ -33,7 +33,7 @@ define([
             }.bind(this));
 
             // Collection Search
-            if (this.$el.hasClass('collectionSearch') && this.ajaxLoad) {
+            if (this.$el.hasClass('collectionSearch') && this.options.ajaxLoad) {
                 // initialize collection search
 
                 $('button[type="submit"]', this.$el).hide();
@@ -44,16 +44,15 @@ define([
                 $('input[name="SearchableText"]', this.$el).on('keyup', function (e) {
                     clearTimeout(delayTimer);
                     // minimum 3 characters before searching
-                    if(($(e.target).val().length > 0) && ($(e.target).val().length < 3)) return;
-                    delayTimer = setTimeout(function() {
+                    if (($(e.target).val().length > 0) && ($(e.target).val().length < 3)) return;
+                    delayTimer = setTimeout(function () {
                         var collectionURL = $(e.target).data('url');
                         var val = encodeURIComponent($(e.target).val());
-                        if(val) {
+                        if (val) {
                             collectionURL += '&' + $(e.target).attr('name') + '=' + val;
                         }
                         $(this.trigger).trigger(
-                            'collectionfilter:reload',
-                            {
+                            'collectionfilter:reload', {
                                 collectionUUID: this.options.collectionUUID,
                                 targetFilterURL: collectionURL,
                                 noReloadSearch: true
@@ -62,7 +61,7 @@ define([
                         this.reloadCollection(collectionURL);
                     }.bind(this), 500);
                 }.bind(this));
-                return;  // no more action neccesarry.
+                return; // no more action neccesarry.
             }
 
             // OPTION 1 - filter rendered as links
@@ -75,8 +74,7 @@ define([
                 var collectionURL = $(e.target).closest('a').attr('href');
 
                 $(this.trigger).trigger(
-                    'collectionfilter:reload',
-                    {
+                    'collectionfilter:reload', {
                         collectionUUID: this.options.collectionUUID,
                         targetFilterURL: collectionURL
                     }
@@ -90,8 +88,7 @@ define([
                 var collectionURL = $(e.target).data('url');
 
                 $(this.trigger).trigger(
-                    'collectionfilter:reload',
-                    {
+                    'collectionfilter:reload', {
                         collectionUUID: this.options.collectionUUID,
                         targetFilterURL: collectionURL
                     }
@@ -107,8 +104,7 @@ define([
                 var collectionURL = $(option).data('url');
 
                 $(this.trigger).trigger(
-                    'collectionfilter:reload',
-                    {
+                    'collectionfilter:reload', {
                         collectionUUID: this.options.collectionUUID,
                         targetFilterURL: collectionURL
                     }
@@ -122,7 +118,7 @@ define([
                 $('.pat-leaflet', this.$el).on('leaflet.moveend leaflet.zoomend', function (e, le) {
                     var narrow_down = $(e.target).data('narrow-down-result');
                     // do nothing if not narrowing down result
-                    if(narrow_down.toLowerCase() === 'false') return;
+                    if (narrow_down.toLowerCase() === 'false') return;
 
                     if (this._initmap_cycles > 0) {
                         // Do not trigger filter when initializing the map.
@@ -133,11 +129,11 @@ define([
 
                     var levent = le['original_event'];
                     // prevent double loading when zooming (because it's always a move too)
-                    if(levent.type === 'moveend' && this._zoomed) {
+                    if (levent.type === 'moveend' && this._zoomed) {
                         this._zoomed = false;
                         return
                     }
-                    if(levent.type === 'zoomend') this._zoomed = true;
+                    if (levent.type === 'zoomend') this._zoomed = true;
                     var collectionURL = $(e.target).data('url'),
                         bounds = levent.target.getBounds();
                     // generate bounds query
@@ -145,8 +141,7 @@ define([
                     collectionURL += "&longitude.query:list:record=" + bounds._northEast.lng + "&longitude.query:list:record=" + bounds._southWest.lng + "&longitude.range:record=minmax";
 
                     $(this.trigger).trigger(
-                        'collectionfilter:reload',
-                        {
+                        'collectionfilter:reload', {
                             collectionUUID: this.options.collectionUUID,
                             targetFilterURL: collectionURL,
                             noReloadMap: true
@@ -159,7 +154,7 @@ define([
         },
 
         reload: function (filterURL) {
-            if (!this.ajaxLoad) {
+            if (!this.options.ajaxLoad) {
                 window.location.href = filterURL;
                 return
             }
@@ -170,21 +165,20 @@ define([
             var query = [].concat(query1, query2).join('&');
             reloadURL = [].concat(urlParts[0], query).join('?');
             var cl = new this.contentloader(
-                this.$el,
-                {
+                this.$el, {
                     url: reloadURL,
-                    target: 'div.portletContent',
-                    content: 'div.portletContent',
+                    target: this.$el,
+                    content: 'aside',
                     trigger: 'immediate'
                 }
             );
         },
 
         reloadCollection: function (collectionURL) {
-            if (!this.ajaxLoad)
+            if (!this.options.ajaxLoad)
                 return;
             var cl = new this.contentloader(
-                $(this.options.contentSelector).parent(),  // let base element for setting classes and triggering events be the parent, which isn't replaced.
+                $(this.options.contentSelector).parent(), // let base element for setting classes and triggering events be the parent, which isn't replaced.
                 {
                     url: collectionURL + '&ajax_load=1',
                     target: this.options.contentSelector,
@@ -203,8 +197,9 @@ define([
             // tiles.
             re = /@@.*\//;
             collectionURL = collectionURL.replace(re, '');
-            window.history.replaceState(
-                {path: collectionURL},
+            window.history.replaceState({
+                    path: collectionURL
+                },
                 '',
                 collectionURL
             );